Refactor : DailyRecord record 도메인으로 이동
