+++ ilkda/src/main/java/com/ilkda/server/record/repository/DailyRecordRepository.java
@@ -1,6 +1,6 @@
-package com.ilkda.server.member.repository;
+package com.ilkda.server.record.repository;
 
-import com.ilkda.server.member.model.DailyRecord;
+import com.ilkda.server.record.model.DailyRecord;
 import org.springframework.data.repository.CrudRepository;
 
 public interface DailyRecordRepository extends CrudRepository<DailyRecord, Long> {
