+++ ilkda/src/main/java/com/ilkda/server/record/service/RecordService.java
@@ -7,9 +7,9 @@ import com.ilkda.server.member.model.Member;
 import com.ilkda.server.member.repository.MemberRepository;
 import com.ilkda.server.record.dto.RecordTextForm;
 import com.ilkda.server.record.dto.RegisterRecordForm;
-import com.ilkda.server.member.model.DailyRecord;
+import com.ilkda.server.record.model.DailyRecord;
 import com.ilkda.server.record.model.Record;
-import com.ilkda.server.member.repository.DailyRecordRepository;
+import com.ilkda.server.record.repository.DailyRecordRepository;
 import com.ilkda.server.record.repository.RecordRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
